% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocketchatr-package.R
\docType{package}
\name{rocketchatr-package}
\alias{rocketchatr}
\alias{rocketchatr-package}
\title{rocketchatr: Rocket.Chat R Client}
\description{
An R client to interact with a Rocket.Chat server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/rocketchatr}
  \item Report bugs at \url{https://github.com/mcanouil/rocketchatr/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{mickael.canouil@cnrs.fr} (\href{https://orcid.org/0000-0002-3396-454}{ORCID})

}
\keyword{internal}
